refactor inner class
