+++ Backend/src/main/java/idealab/api/dto/helper/PrintJobColorCount.java
@@ -1,0 +1,30 @@
+package idealab.api.dto.helper;
+
+public class PrintJobColorCount {
+    private String color;
+    private Integer count;
+
+    public String getColor() {
+        return color;
+    }
+
+    public void setColor(String color) {
+        this.color = color;
+    }
+
+    public Integer getCount() {
+        return count;
+    }
+
+    public void setCount(Integer count) {
+        this.count = count;
+    }
+
+    @Override
+    public String toString() {
+        return "PrintJobColorCount{" +
+                "color='" + color + '\'' +
+                ", count=" + count +
+                '}';
+    }
+}
