+++ Backend/src/main/java/idealab/api/dto/response/AuditPrintJobColorCountResponse.java
@@ -1,19 +1,19 @@
 package idealab.api.dto.response;
 
-import idealab.api.operations.AuditOperations;
+import idealab.api.dto.helper.PrintJobColorCount;
 
 import java.util.List;
 import java.util.Map;
 
 public class AuditPrintJobColorCountResponse extends GenericResponse {
 
-    private Map<Integer, List<AuditOperations.PrintJobColorCount>> data;
+    private Map<Integer, List<PrintJobColorCount>> data;
 
-    public Map<Integer, List<AuditOperations.PrintJobColorCount>> getData() {
+    public Map<Integer, List<PrintJobColorCount>> getData() {
         return data;
     }
 
-    public void setData(Map<Integer, List<AuditOperations.PrintJobColorCount>> data) {
+    public void setData(Map<Integer, List<PrintJobColorCount>> data) {
         this.data = data;
     }
